refactor(api): restructure importer/api/ subpackage [T-09] (#41)
Split flat api.py / auth_api.py / users_api.py into importer/api/
subpackage with focused modules: books, auth, users, config,
import_pipeline. Delete dead views.py and forms.py. Extract
serialize_book() as a shared function. Fix logger.warn → logger.warning,
bare except: → except Exception, Setting.objects.first() → Setting.load(),
and remove late import of rootdir from the deleted views module.

Closes #15
